rearrange powerlaw to support initialization hooks
