+++ networks/powerlaw/Make.package
@@ -3,10 +3,13 @@ F90EXE_sources += network_properties.F90
 
 CEXE_headers += network_properties.H
 CEXE_headers += actual_network.H
+CEXE_sources += actual_network_data.cpp
 
 ifeq ($(USE_REACT),TRUE)
 ifneq ($(USE_SIMPLIFIED_SDC), TRUE)
 F90EXE_sources += actual_burner.F90
 endif
 F90EXE_sources += actual_rhs.F90
+CEXE_headers += actual_rhs.H
+CEXE_sources += actual_rhs_data.cpp
 endif
