+++ networks/powerlaw/actual_network_data.cpp
@@ -1,0 +1,3 @@
+#include <actual_network.H>
+
+void actual_network_init() {}
