+++ networks/powerlaw/actual_rhs_data.cpp
@@ -1,0 +1,3 @@
+#include <actual_rhs.H>
+
+void actual_rhs_init() {}
